Added a command line tool to process batches of addresses as needed.
The tool can be called by navigating into the usps-service directory (the repository itself) and invoking
python usps_service. The tool takes a variety of arguments, which can be viewed using the -h flag. The input
addresses should be in a csv file where the first line is populated with the column headers. All headers must
be unique. The pieces of the address (street, city, state, zip) must be in distinct columns.

The tool processes each address in the file and produces output containing both the contents of the original
csv and the results from USPS service. That output is placed in a new csv file in the same directory as the
original.