Set up Jenkinsfile to allow Jenkins to find and automatically test all pushed commits. (#4)
Created two .sh files to run the prebuild process and the test process. These allow us to invoke source
activate (for the virtual environment) using bash. During prebuild, we create a temp directory stored in
~/temp for the virtual environment itself. The name of the temp directory is stored in ~/temp/venv. We
create the virtual environment and install the requirements.

The test script simply reactivates the virtual environment and runs all test files in the test directory.